pyproject.toml
[project]
name = "documentgenerationanek"
version = "0.1.0"
description = "Lean document generator backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "fastapi>=0.116.0",
    "minio>=7.2.16",
    "pydantic-settings>=2.10.1",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.20",
    "reportlab>=4.4.3",
    "sqlalchemy>=2.0.41",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "ruff>=0.12.0",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"app/models/*.py" = ["F821"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: API-level integration tests",
    "migration: Alembic migration validation tests",
]

[tool.mypy]
python_version = "3.12"
files = ["app", "tests"]
exclude = [
    "^app/models/",
    "^app/repositories/",
    "^migrations/",
]
ignore_missing_imports = true
check_untyped_defs = true
warn_unused_ignores = true
pretty = true

[[tool.mypy.overrides]]
module = [
    "app.models.*",
    "app.services.generation.document_composer_service",
]
ignore_errors = true
